requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
dj-database-url==0.5.0
Django==4.0.5
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==21.2
django-environ==0.8.1
django-extensions==3.1.5
django-on-heroku==1.1.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydot==1.4.2
pydotplus==2.0.2
pyparsing==3.0.8
python-decouple==3.6
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
whitenoise==6.0.0
